Add custom serialiser using the Symfony serializer. Get blog post via groups
