+++ src/Service/CustomSerialiserInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Service;
+
+interface CustomSerialiserInterface
+{
+    public function serialise(object|array $data, array $groups = []): string;
+}
